+++ tests/test_integrated_channels/test_degreed/test_exporters/test_learner_data.py
@@ -32,6 +32,7 @@ class TestDegreedLearnerExporter(unittest.TestCase):
     def setUp(self):
         self.user = factories.UserFactory(username='C3PO', id=1, email='degreed@email.com')
         self.course_id = 'course-v1:edX+DemoX+DemoCourse'
+        self.course_key = 'edX+DemoX'
         self.enterprise_customer = factories.EnterpriseCustomerFactory()
         self.enterprise_customer_user = factories.EnterpriseCustomerUserFactory(
             user_id=self.user.id,
@@ -78,15 +79,17 @@ class TestDegreedLearnerExporter(unittest.TestCase):
             completed_date=completed_date,
             is_passing=is_passing,
         )
-        learner_data_record = learner_data_records[0]
+        assert len(learner_data_records) == 2
+        assert learner_data_records[0].course_id == self.course_key
+        assert learner_data_records[1].course_id == self.course_id
 
-        assert learner_data_record.enterprise_course_enrollment_id == enterprise_course_enrollment.id
-        assert learner_data_record.degreed_user_email == 'degreed@email.com'
-        assert learner_data_record.course_id == enterprise_course_enrollment.course_id
-        assert learner_data_record.course_completed == (completed_date is not None and is_passing)
-        assert learner_data_record.completed_timestamp == (
-            self.NOW.strftime('%F') if completed_date is not None else None
-        )
+        for learner_data_record in learner_data_records:
+            assert learner_data_record.enterprise_course_enrollment_id == enterprise_course_enrollment.id
+            assert learner_data_record.degreed_user_email == 'degreed@email.com'
+            assert learner_data_record.course_completed == (completed_date is not None and is_passing)
+            assert learner_data_record.completed_timestamp == (
+                self.NOW.strftime('%F') if completed_date is not None else None
+            )
 
     def test_no_remote_id(self):
         """