+++ CHANGELOG.rst
@@ -14,6 +14,21 @@ Change Log
 Unreleased
 ----------
 
+[0.70.7] - 2018-07-12
+---------------------
+
+* Make customer catalog content filter's default value configurable.
+
+[0.70.6] - 2018-07-09
+---------------------
+
+* Pass catalog value only when provided on enterprise course enrollment page.
+
+[0.70.5] - 2018-07-06
+---------------------
+
+* Send learner data transmissions to integrated channels by course key and course run id.
+
 [0.70.4] - 2018-07-03
 ---------------------